Add due date, repeating tasks, and paused tasks
This patch adds dated tasks. A dated task has a due date specified
with "due:yyyy-mm-dd" in the description. A repeating task has a
due date and "rep:" with a repeat specification. A repeat
specification is one of the following forms:

    1d      one day
    2w      two weeks
    3m      three months
    4y      four years
    1m;1d   one month with lead time one day
    ~1w     approximately one week

Approximately timed tasks always have a lead time of one day, and
marking them completed calculates the next due date from the current
date, not the due date.

Tasks may be paused with "pause:1" in the description.